Name fixes for four Frisian places and province
This uses the current names of four Frisian villages as used in Dutch
(and Frisian), which have in the past half century supplanted their
older Dutch forms. The villages are Jirnsum, Grou, Eastermar, and De
Westereen.

For the province of Frisia its English exonym is used, and in Dutch the
official name. The Dutch vernacular name is kept in `nld_x_colloquial`.

Some remaining `dut_*` keys were removed in favour of their `nld_*`
companions.